♻ Refactoring of the code - Change of the project's file structure
