+++ src/App.tsx
@@ -1,15 +1,12 @@
 import { MantineProvider } from "@mantine/core";
 import "./App.css";
 import Planner from "./components/planner/Planner";
-import ReadExcelFile from "./components/ReadExcelFile";
-import theme from "./utils/style/MantineTheme";
+import ReadExcelFile from "./components/DataGrid/ReadExcelFile";
+import theme from "./assets/style/MantineTheme";
 
 function App() {
   return (
     <MantineProvider withGlobalStyles withNormalizeCSS theme={theme}>
-      {/* {[...Array(10)].map(() => (
-            <LibelleCard />
-          ))} */}
       <Planner />
       {/* <ReadExcelFile /> */}
     </MantineProvider>
