+++ src/data/interfaces/IRessource.ts
